+++ features/ftp_server/step_definitions/debug.rb
@@ -1,8 +1,0 @@
-Then /^the server should have written( no)? debug output$/ do |neg|
-  method = if neg
-             :should
-           else
-             :should_not
-           end
-  server.debug_output.send(method) == ''
-end