+++ features/ftp_server/step_definitions/logging.rb
@@ -1,0 +1,8 @@
+Then /^the server should have written( no)? log output$/ do |neg|
+  method = if neg
+             :should
+           else
+             :should_not
+           end
+  server.log_output.send(method) == ''
+end