+++ lib/ftpd.rb
@@ -1,4 +1,5 @@
 require 'fileutils'
+require 'logger'
 require 'memoizer'
 require 'openssl'
 require 'pathname'
@@ -20,6 +21,7 @@ module Ftpd
   autoload :FileSystemMethodMissing,   'ftpd/file_system_method_missing'
   autoload :FtpServer,                 'ftpd/ftp_server'
   autoload :InsecureCertificate,       'ftpd/insecure_certificate'
+  autoload :NullLogger,                'ftpd/null_logger'
   autoload :ReadOnlyDiskFileSystem,    'ftpd/read_only_disk_file_system'
   autoload :Server,                    'ftpd/server'
   autoload :Session,                   'ftpd/session'