+++ lib/ftpd/ftp_server.rb
@@ -6,20 +6,6 @@ module Ftpd
     DEFAULT_SERVER_NAME = 'wconrad/ftpd'
     DEFAULT_SESSION_TIMEOUT = 300 # seconds
 
-    # If truthy, emit debug information (such as replies received and
-    # responses sent) to the file named by #debug_path.
-    #
-    # Change to this attribute only take effect for new sessions.
-
-    attr_accessor :debug
-
-    # The path to which to write debug information.  Defaults to
-    # '/dev/stdout'
-    #
-    # Change to this attribute only take effect for new sessions.
-
-    attr_accessor :debug_path
-
     # The number of seconds to delay before replying.  This is for
     # testing, when you need to test, for example, client timeouts.
     # Defaults to 0 (no delay).
@@ -60,6 +46,11 @@ module Ftpd
 
     attr_accessor :server_version
 
+    # The logger.  Defaults to nil (no logging).
+    # @return [Logger]
+
+    attr_accessor :log
+
     # Create a new FTP server.  The server won't start until the
     # #start method is called.
     #
@@ -73,14 +64,13 @@ module Ftpd
     def initialize(driver)
       super()
       @driver = driver
-      @debug_path = '/dev/stdout'
-      @debug = false
       @response_delay = 0
       @list_formatter = ListFormat::Ls
       @auth_level = AUTH_PASSWORD
       @session_timeout = 300
       @server_name = DEFAULT_SERVER_NAME
       @server_version = read_version_file
+      @log = nil
     end
 
     private
@@ -88,15 +78,14 @@ module Ftpd
     def session(socket)
       Session.new(:socket => socket,
                   :driver => @driver,
-                  :debug => @debug,
-                  :debug_path => debug_path,
                   :list_formatter => @list_formatter,
                   :response_delay => response_delay,
                   :tls => @tls,
                   :auth_level => @auth_level,
                   :session_timeout => @session_timeout,
                   :server_name => @server_name,
-                  :server_version => @server_version).run
+                  :server_version => @server_version,
+                  :log => log).run
     end
 
     def read_version_file